==> Dolgos_PDOs_Early-Song_Tissue-Song_PDOs/1_harmonization.Rmd <==
---
title: "1.Dolgos_2024 (Early_Passage PDOs) & Song_2022 (Tumor Tissue + Tumor Org): harmonization"
author: "Romuald Parmentier"
date: "2024-11-22"
output: html_document
---

```{r Load libraries and create output folder, message = F}

# Libraries
library(SingleCellExperiment)
library(dplyr)

# Functions and palettes
source(paste0(github_dir, "/Dolgos_Custom_Functions.R"))
source(paste0(github_dir, "/Dolgos_Custom_Color_Palettes.R"))

# Create output path
out_path = create_exp_folder(
  github_dir = github_dir,
  samples_ID = "Dolgos_PDOs_Early-Song_Tissue-Song_PDOs",
  exp = "1_harmonization"
)

```

# Prepare the data

```{r Load all sce annotated}

# Song_2022 Tumor Tissue Epithelial cells only
##############################################

file_path = get_exp_file_path(
  github_dir = github_dir,
  samples_ID = "Song_Tissue",
  prev_exp = "1bis_normalization_rescaling", #1bis = published object pipeline
  pattern = "list_sce_normalized.rds")

Song_Tumor_Tissue_list_sce = readRDS(file_path)

# Load the annotated object to retrieve Malignancy
file_path = get_exp_file_path(
  github_dir = github_dir,
  samples_ID = "Song_Tissue",
  prev_exp = "2bis_dimension_reduction", #2bis = published object pipeline
  pattern = "sce_comb_dim_reduced.rds")

Song_Tumor_Tissue_sce_comb = readRDS(file_path)

# Song_2022 Tumor-Derived Organoids
###################################

file_path = get_exp_file_path(
  github_dir = github_dir,
  samples_ID = "Song_PDOs",
  prev_exp = "2_normalization_rescaling", 
  pattern = "list_sce_normalized.rds")

Song_Tumor_Org_list_sce = readRDS(file_path)
Song_Tumor_Org_list_sce <- Song_Tumor_Org_list_sce[grepl("tumor_derived_org", names(Song_Tumor_Org_list_sce))]

# Load the annotated object to retrieve Malignancy
file_path = get_exp_file_path(
  github_dir = github_dir,
  samples_ID = "Atlas_PCa_PDOs",
  prev_exp = "cell_type_annotation", 
  pattern = "sce_comb_malignancy_annotated.rds")

Song_Tumor_Org_sce_comb = readRDS(file_path)
Song_Tumor_Org_sce_comb = Song_Tumor_Org_sce_comb[ , grepl(x = Song_Tumor_Org_sce_comb$Sample_Description, pattern = "tumor_derived_org")]

# Dolgos_2024_Early_Passage
###########################

file_path = get_exp_file_path(
  github_dir = github_dir,
  samples_ID = "Dolgos_Early",
  prev_exp = "1_normalization_rescaling", 
  pattern = "list_sce_normalized.rds")

Dolgos_list_sce = readRDS(file_path)

# Load the annotated object to retrieve Malignancy
file_path = get_exp_file_path(
  github_dir = github_dir,
  samples_ID = "Dolgos_Early",
  prev_exp = "4_clustering", 
  pattern = "sce_comb_clustered")

Dolgos_sce_comb = readRDS(file_path)

```

```{r Harmonizing Song Tumor Tissue dataset}

# Rename cell types to match ours
Song_Tumor_Tissue_sce_comb$ID = ifelse(
  test = Song_Tumor_Tissue_sce_comb$ID == "LE",
  yes = "Luminal",
  no = ifelse(
    test = Song_Tumor_Tissue_sce_comb$ID == "BE",
    yes = "Basal",
    no = Song_Tumor_Tissue_sce_comb$ID))

sce_id = 1 

for(sce in Song_Tumor_Tissue_list_sce){
  
  sce = sce[, which(colnames(sce) %in% colnames(Song_Tumor_Tissue_sce_comb))]
  
  # Get indice of cells in sce_comb that are in this sce
  ind_cells_in_sce_comb = which(colnames(Song_Tumor_Tissue_sce_comb) %in% colnames(sce))
  
  # Ue this indice to assign cell_Type
  sce$Cell_Type = Song_Tumor_Tissue_sce_comb$ID[ind_cells_in_sce_comb]
  sce$Sample_Type = "Parental_Tumor"
  sce$Dataset = "Song_Tumor_Tissue_2022"

  # Keep only useful colData columns
  colData(sce) = colData(sce)[c("Sample_Description","Dataset","Sample_Type", "Cell_Type" ,"sizeFactor")]
  
  Song_Tumor_Tissue_list_sce[[sce_id]] = sce
  
  sce_id = sce_id + 1
  
}

```

```{r Harmonizing Song Org dataset}

sce_id = 1 

for(sce in Song_Tumor_Org_list_sce){
  
  sce = sce[, which(colnames(sce) %in% colnames(Song_Tumor_Org_sce_comb))]
  
  # Get indice of cells in sce_comb that are in this sce
  ind_cells_in_sce_comb = which(colnames(Song_Tumor_Org_sce_comb) %in% colnames(sce))
  
  # Ue this indice to assign cell_Type
  sce$Cell_Type = Song_Tumor_Org_sce_comb$Malignancy[ind_cells_in_sce_comb]
  sce$Sample_Type = "Matrigel_PDOs"
  sce$Dataset = "Song_Organoids_2022"

  # Keep only useful colData columns
  colData(sce) = colData(sce)[c("Sample_Description","Dataset", "Sample_Type", "Cell_Type" ,"sizeFactor")]
  
  Song_Tumor_Org_list_sce[[sce_id]] = sce
  
  sce_id = sce_id + 1
  
}

```


```{r Harmonizing Dolgos_Early_Passage dataset}

sce_id = 1 

for(sce in Dolgos_list_sce){
  
  sce = sce[, which(colnames(sce) %in% colnames(Dolgos_sce_comb))]
  
  # Get indice of cells in sce_comb that are in this sce
  ind_sce_comb_cells = which(colnames(Dolgos_sce_comb) %in% colnames(sce))
  
  # Use this indice to assign cell_Type
  sce$Cell_Type = Dolgos_sce_comb$Malignancy[ind_sce_comb_cells]
  
  sce$Sample_Type <- ifelse(
    test = grepl(x = sce$Sample_Description, pattern = "Org_ECMf"),
    yes = "ECMf_PDOs",
    no = "Matrigel_PDOs")
  
  sce$Dataset = "Dolgos_2024_Early_Passage"
  
  # Keep only useful colData columns
  colData(sce) = colData(sce)[c("Sample_Description","Dataset","Sample_Type", "Cell_Type", "sizeFactor")]
  
  Dolgos_list_sce[[sce_id]] = sce
  
  sce_id = sce_id + 1
  
}

```

```{r Combine all sce in a single list}

# Gather all sce_comb in a single list
list_all_sce = list(Dolgos_list_sce, Song_Tumor_Org_list_sce, Song_Tumor_Tissue_list_sce)
list_all_sce <- do.call(c, list_all_sce)

```

```{r Harmonizing count matrix ronames and colnames}

##############################
### rownames (= features) ###
#############################

list_all_sce = restrict_reorder_features_names(sce_list = list_all_sce)

##################################
### colnames (= cell barcodes) ###
##################################

# Check colnames of the matrix (normally done in previous steps)
lapply(list_all_sce, function(x){head(colnames(x))})

```

```{r Harmonizing rowData and coldata}

###############
### RowData ###
###############

# Identify the common rowData columns across all SCE objects
common_columns <- Reduce(intersect, lapply(list_all_sce, function(sce) colnames(rowData(sce))))

# Update each SCE object to retain only the common rowData columns
list_all_sce <- lapply(list_all_sce, function(sce) {
  rowData(sce) <- rowData(sce)[, common_columns, drop = FALSE]  # Subset rowData
  return(sce)  # Return updated sce
})

# Check colnames of the rowData (performed above)
lapply(list_all_sce, function(x){colnames(rowData(x))})

###############
## ColData ###
##############

# Check colnames of the colData (normally done in previous steps)
lapply(list_all_sce, function(x){colnames(colData(x))})


```

```{r Creating a sce_comb object with all sce}

## Get total raw counts, first in a list, then combine them into a matrix
counts_comb = lapply(list_all_sce, function(x) {counts(x)})
counts_comb <- Reduce(function(x, y) cbind(x, y), counts_comb)

## Then get total logNormed counts, first in a list, then combine them into a matrix
logcounts_comb = lapply(list_all_sce, function(x) {logcounts(x)})
logcounts_comb <- Reduce(function(x, y) cbind(x, y), logcounts_comb)

# Combine all cell metadata (columns are already the same for each object of the list)
colData_comb = lapply(list_all_sce, function(x) {colData(x)})
colData_comb = Reduce(function(x, y) rbind(x, y), colData_comb)

# Combine the slots to create the combined sce object
sce_comb <- SingleCellExperiment( 
  assays = list(counts = counts_comb, logcounts = logcounts_comb),  
  rowData = rowData(list_all_sce[[1]]), # RowData is similar among all elements 
  colData = colData_comb
) 

```

```{r Export list_sce and sce_comb}

###########
# Export list_sce_comb object
##########

saveRDS(
  object = list_all_sce,
  file = paste0(out_path,time_stamp(),"list_all_sce_comb_harmonized.rds"))

saveRDS(
  object = sce_comb,
  file = paste0(out_path,time_stamp(),"sce_comb_all_harmonized.rds"))

```
