==> Dolgos_PDOs_Early-Song_Tissue-Song_PDOs/5_signature_score.Rmd <==
---
title: "Dolgos_2024_Early_Passage integration with Song Tum Tissue and Tum Org: gene signature score"
author: "Romuald Parmentier"
date: "2024-11-28"
output: html_document
---

```{r Load libraries and create output folder, message = F}

# Libraries
library(Seurat)
library(SingleCellExperiment)
library(ggplot2)
library(dplyr)
library(tidyr)
library(msigdbr)
library(org.Hs.eg.db)
library(clusterProfiler)
library(AnnotationHub)

# Functions and palettes
source(paste0(github_dir, "/Dolgos_Custom_Functions.R"))
source(paste0(github_dir, "/Dolgos_Custom_Color_Palettes.R"))


# Create output path
out_path = create_exp_folder(
  github_dir = github_dir,
  samples_ID = "Dolgos_PDOs_Early-Song_Tissue-Song_PDOs",
  exp = "5_signature_score"
)

```

# Prepare the data

```{r Load the files}

# Load sce after nornalization (no rescaling)
file_path = get_exp_file_path(
  github_dir = github_dir,
  samples_ID = "Dolgos_PDOs_Early-Song_Tissue-Song_PDOs",
  prev_exp = "3_dimension_reduction", 
  pattern = "sce_comb_dim_reduced.rds") # Sce after batch correction for visuaoization

sce_comb = readRDS(file = file_path)

```

```{r Add a category to sce_comb }

sce_comb$Dataset_Cell_Type = paste0(
  sub("_.*", "", sce_comb$Dataset), 
  "_", sub("_.*", "", sce_comb$Sample_Type),
  "_", sce_comb$Cell_Type)

```

```{r Transform sce to Seurat object}

seurat_comb = as.Seurat(
  x = sce_comb, 
  counts = "counts", 
  data = "logcounts"
)

# Set the originalexp (aka "RNA" as the default assay)
DefaultAssay(seurat_comb) = "originalexp"

# Adds scale.data slot
seurat_comb = ScaleData(seurat_comb)

```

# Perform signature score

```{r A}

# Define the list of panels
panels = c(
  "C5" = "GOBP_GRANULOCYTE_CHEMOTAXIS",
  "C5" = "GOBP_NEUTROPHIL_CHEMOTAXIS",
  "C5" = "GOBP_HUMORAL_IMMUNE_RESPONSE",
  "C5" = "GOBP_LEUKOCYTE_CHEMOTAXIS",
  "C5" = "GOBP_LEUKOCYTE_MIGRATION",
  "C5" = "GOBP_RESPONSE_TO_CHEMOKINE",
  "C5" = "GOBP_RESPONSE_TO_TUMOR_NECROSIS_FACTOR",
  "C5" = "GOBP_CELLULAR_SENESCENCE",
  "C8" = "TRAVAGLINI_LUNG_CLUB_CELL",
  "H" = "HALLMARK_IL6_JAK_STAT3_SIGNALING"
)

# Iterate over each panel in the list
for (panel_name in panels) {
  
  # Make a table with all genes belonging to the signature of interest
  gene_panel = get_signature_data(
    pathway_proxy = panel_name, MSigDB_category = names(panel_name)
  )
  
  # Calculate signature score and add it to Seurat object as an alternative assay
  seurat_comb = calculate_signature_score(
    assay_name = "originalexp",
    seurat_object = seurat_comb, 
    signature_name = panel_name, 
    signature_genes = gene_panel$SYMBOL
  )
  
  # Transfer the signature score from Seurat to SCE object
  sce_comb = add_signature_score_to_sce(
    sce_object = sce_comb,
    seurat_object = seurat_comb,
    signature_name = panel_name,
    seurat_assay = paste0(panel_name, "_score_only"),
    signature_genes = gene_panel$SYMBOL
  )
  
  # Calculate min and max signature score outliers for better visualization on UMAP
  min_max_outliers = get_signature_score_outliers( 
    sce_object = sce_comb,
    sce_signature_column = paste0(panel_name, "_score"),
    signature_name = panel_name,
    nb_outliers = 1
  )
  
}

```

# Visualizing signature scores

```{r Make a tibble of metadata}

# get metadata
md <- as_tibble(colData(sce_comb))

# get embeddings from 2 first PC
coords <- as_tibble(reducedDim(sce_comb, "UMAP_on_MNN.1-20"))[,1:2]
colnames(coords) = c("UMAP_1","UMAP_2")

# combine dataframes
md <- cbind(md, coords)

```

```{r Violin plot of signature}

# Define the list of panels
panels = c(
  "C5" = "GOBP_GRANULOCYTE_CHEMOTAXIS_score",
  "C5" = "GOBP_NEUTROPHIL_CHEMOTAXIS_score",
  "C5" = "GOBP_HUMORAL_IMMUNE_RESPONSE_score",
  "C5" = "GOBP_LEUKOCYTE_CHEMOTAXIS_score",
  "C5" = "GOBP_LEUKOCYTE_MIGRATION_score",
  "C5" = "GOBP_RESPONSE_TO_CHEMOKINE_score",
  "C5" = "GOBP_RESPONSE_TO_TUMOR_NECROSIS_FACTOR_score",
  "C5" = "GOBP_CELLULAR_SENESCENCE_score",
  "C8" = "TRAVAGLINI_LUNG_CLUB_CELL_score",
  "H" = "HALLMARK_IL6_JAK_STAT3_SIGNALING_score"
)

md_club = md %>%
  dplyr::filter(grepl(x = Dataset_Cell_Type , pattern = "Club"))

# Iterate over each panel in the list
for (panel_name in panels) {
  
  plot <- ggplot(md_club, aes(x = Dataset_Cell_Type, y = .data[[panel_name]], fill = Dataset_Cell_Type)) +
    geom_violin(alpha = 0.8, position = position_dodge(width = 0.8), scale = "width") +
    geom_boxplot(width = 0.1, position = position_dodge(width = 0.8), alpha = 0.5) +
    guides(fill = "none") +
    theme_minimal() + 
    theme(
      axis.text.x = element_text(angle = 45, hjust = 1),
      axis.title = element_text(size = 12),  # Customize axis title appearance
      axis.text = element_text(size = 10),    # Customize axis text appearance
      axis.line = element_line(size = 0.5),   # Customize axis line appearance
      axis.ticks = element_line(size = 0.5)) +
    theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
    labs(
      title = element_text(panel_name),
      x = element_blank(),
      y = "Signature score")
  
  ggsave(
    plot = plot,
    filename = paste0(out_path,time_stamp(),"plots_Violin_Scores_", panel_name,".pdf"),
    device = "pdf",
    width = 210,
    height = 150,
    units = "mm"
  )
  
}

```

```{r Single UMAP colored signature score}

panels = c(
  "C5" = "GOBP_GRANULOCYTE_CHEMOTAXIS",
  "C5" = "GOBP_NEUTROPHIL_CHEMOTAXIS",
  "C5" = "GOBP_HUMORAL_IMMUNE_RESPONSE",
  "C5" = "GOBP_LEUKOCYTE_CHEMOTAXIS",
  "C5" = "GOBP_LEUKOCYTE_MIGRATION",
  "C5" = "GOBP_RESPONSE_TO_CHEMOKINE",
  "C5" = "GOBP_RESPONSE_TO_TUMOR_NECROSIS_FACTOR",
  "C5" = "GOBP_RESPONSE_TO_TYPE_II_INTERFERON",
  "C5" = "GOBP_CELLULAR_SENESCENCE",
  "C8" = "TRAVAGLINI_LUNG_CLUB_CELL",
  "H" = "HALLMARK_IL6_JAK_STAT3_SIGNALING"
)

# Iterate over each panel in the list
for (panel_name in panels) {
  
  plot <- ggplot() +
    # Rasterise the first layer
    ggrastr::rasterise(
      geom_point(
        data = md,
        aes(x = UMAP_1, y = UMAP_2, fill = .data[[panel_name]]), 
        alpha = 0.7, size = 1.3, 
        shape = 21,
        color = "gray10",
        stroke = 0.2),  # Lighter stroke color (adjust as needed)),# Ensure shape supports fill and color
      dpi = 300, scale = 1) +
    # Create a continuous gradient for the fill color
    scale_fill_gradientn(colors = c("lightgrey", "lightgrey", "#FF7078FF", "#69000CFF"),  # Grey to red gradient
                         values = scales::rescale(c(-0.15,0.20, 0.35, 1))) +
    theme_classic() + # Set limits
    labs(title = panel_name,
         fill = "Score") +  # Change the legend title
    guides(
      alpha = "none",
      size = "none"
    )
  
  ggsave(
    plot = plot,
    filename = paste0(out_path,time_stamp(),"plot_UMAP_", panel_name, ".pdf"),
    device = "pdf",
    width = 10,
    height = 8,
  )
  
}

```

```{r}

sce_comb_club = sce_comb[,grepl("Club", sce_comb$Dataset_Cell_Type)]

# Iterate over each panel in the list
for (panel_name in panels) {
  
  panel <- paste0(panel_name,"_gene")
  
  # Check if the genes exist in the expression matrix
  available_genes <- rowData(sce_comb_club)[[panel]]
  
  HeatMap_signature_genes(
    sce_object = sce_comb_club,
    genes = rownames(sce_comb_club)[available_genes],
    signature_name = "GOBP_HUMORAL_IMMUNE_RESPONSE", 
    group1 = "Dataset_Cell_Type",
    group1_color = pal_dataset_cell_type,
    group2 = NULL,
    group2_color = NULL,
    gene_text_size  = 5, 
    scale_color = c("lightgrey",rev(paletteer_c("viridis::magma", 20))),
    scale = F,
    out_path = out_path)
  
}

```

# Export files

```{r Export sce object with signature scores }

# Clustering takes a lot of time, better to export sce object here

saveRDS(
  object = sce_comb,
  file = paste0(out_path,time_stamp(),"sce_comb_signature_score.rds")
)

```
